requirements.txt
peewee==3.16.0
psycopg2==2.9.5
flask==2.2.3
python-dotenv==1.0.0
pika==1.3.1
